refactor: move data-table + editable-trigger + data-import into blocks/data-table
Reorganizes related modules under a single `apps/web/src/blocks/data-table/`
namespace. `data-table-v2/` → `blocks/data-table/`, `components/editable-trigger/`
→ `blocks/data-table/editable-trigger/`, `features/data-import/` →
`blocks/data-table/data-import/`. All imports updated.
