+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/-contacts/contact-transacoes-tab.tsx
@@ -15,12 +15,12 @@ import {
 import { Receipt } from "lucide-react";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { toast } from "sonner";
-import { DataTableBody } from "@/components/data-table-v2/data-table-body";
-import { DataTableContainer } from "@/components/data-table-v2/data-table-container";
-import { DataTableEmptyState } from "@/components/data-table-v2/data-table-empty-state";
-import { DataTableHeader } from "@/components/data-table-v2/data-table-header";
-import { DataTableRoot } from "@/components/data-table-v2/data-table-root";
-import { useDataTableLayout } from "@/components/data-table-v2/use-data-table-layout";
+import { DataTableBody } from "@/blocks/data-table/data-table-body";
+import { DataTableContainer } from "@/blocks/data-table/data-table-container";
+import { DataTableEmptyState } from "@/blocks/data-table/data-table-empty-state";
+import { DataTableHeader } from "@/blocks/data-table/data-table-header";
+import { DataTableRoot } from "@/blocks/data-table/data-table-root";
+import { useDataTableLayout } from "@/blocks/data-table/use-data-table-layout";
 import { useSheet } from "@/hooks/use-sheet";
 import type { Outputs } from "@/integrations/orpc/client";
 import { orpc } from "@/integrations/orpc/client";
