+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/-transactions/transactions-list.tsx
@@ -46,24 +46,24 @@ import {
 } from "lucide-react";
 import { startTransition, useCallback, useMemo, useState } from "react";
 import { toast } from "sonner";
-import { DataTableBody } from "@/components/data-table-v2/data-table-body";
-import { DataTableBulkActionBar } from "@/components/data-table-v2/data-table-bulk-action-bar";
-import { DataTableColumnVisibility } from "@/components/data-table-v2/data-table-column-visibility";
-import { DataTableContainer } from "@/components/data-table-v2/data-table-container";
-import { DataTableEmptyState } from "@/components/data-table-v2/data-table-empty-state";
-import { DataTableHeader } from "@/components/data-table-v2/data-table-header";
-import { DataTablePagination } from "@/components/data-table-v2/data-table-pagination";
-import { DataTableRoot } from "@/components/data-table-v2/data-table-root";
-import { DataTableSearch } from "@/components/data-table-v2/data-table-search";
+import { DataTableBody } from "@/blocks/data-table/data-table-body";
+import { DataTableBulkActionBar } from "@/blocks/data-table/data-table-bulk-action-bar";
+import { DataTableColumnVisibility } from "@/blocks/data-table/data-table-column-visibility";
+import { DataTableContainer } from "@/blocks/data-table/data-table-container";
+import { DataTableEmptyState } from "@/blocks/data-table/data-table-empty-state";
+import { DataTableHeader } from "@/blocks/data-table/data-table-header";
+import { DataTablePagination } from "@/blocks/data-table/data-table-pagination";
+import { DataTableRoot } from "@/blocks/data-table/data-table-root";
+import { DataTableSearch } from "@/blocks/data-table/data-table-search";
 import {
    DataTableToolbar,
    DataTableToolbarGroup,
-} from "@/components/data-table-v2/data-table-toolbar";
-import { useDataTableLayout } from "@/components/data-table-v2/use-data-table-layout";
-import { DataImportButton } from "@/features/data-import/data-import-button";
-import { DataImportSection } from "@/features/data-import/data-import-section";
-import { useDataImport } from "@/features/data-import/use-data-import";
-import type { DataImportConfig } from "@/features/data-import/types";
+} from "@/blocks/data-table/data-table-toolbar";
+import { useDataTableLayout } from "@/blocks/data-table/use-data-table-layout";
+import { DataImportButton } from "@/blocks/data-table/data-import/data-import-button";
+import { DataImportSection } from "@/blocks/data-table/data-import/data-import-section";
+import { useDataImport } from "@/blocks/data-table/data-import/use-data-import";
+import type { DataImportConfig } from "@/blocks/data-table/data-import/types";
 import { PageFilters } from "@/components/page-filters/page-filters";
 import { PageFilter } from "@/components/page-filters/page-filter";
 import { useSheet } from "@/hooks/use-sheet";
