+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/transactions.tsx
@@ -3,7 +3,7 @@ import { createFileRoute } from "@tanstack/react-router";
 import { useCallback } from "react";
 import { z } from "zod";
 import { DefaultHeader } from "../-layout/default-header";
-import { DataTableSkeleton } from "@/components/data-table-v2/data-table-skeleton";
+import { DataTableSkeleton } from "@/blocks/data-table/data-table-skeleton";
 import { QueryBoundary } from "@/components/query-boundary";
 import { TransactionsList } from "./-transactions/transactions-list";
 import { buildTransactionColumns } from "./-transactions/transactions-columns";
